BUGFIX: ensure in cloud environments, only one content release is building at any given time
We usually rely on prunner to ensure that only one build is running at any given time.

However, when running in a cloud environment with no shared storage, the prunner data
folder is not shared between instances. In this case, during a deployment, two containers
run concurrently, with two separate prunner instances (the old and the new one), which
do not see each other.

We could fix this in prunner itself, but this would be a bigger undertaking
(different storage backends for prunner), or we can work around this in
DecoupledContentStore. This is what we do here.